refactor: move new mantine components in dedicated folder + split home
