+++ inertia/mantine/layouts/_mantine_base_layout.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from 'react-i18next';
 
 import '@mantine/core/styles.css';
 import '@mantine/spotlight/styles.css';
-import '../../../styles/index.css';
+import '../../styles/index.css';
 
 const theme = createTheme({});
 const TRANSITION_IN_CLASS = '__transition_fadeIn';
