+++ inertia/pages/terms.tsx
@@ -2,7 +2,7 @@ import { Link } from '@inertiajs/react';
 import { route } from '@izzyjs/route/client';
 import { ReactNode } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
-import { MantineContentLayout } from '~/components/layouts/mantine/mantine_content_layout';
+import { MantineContentLayout } from '~/mantine/layouts/mantine_content_layout';
 
 function TermsPage() {
 	const { t } = useTranslation('terms');
